feat: add QR code functionality to challenge voting page
Adds the Endroid QR Code library to enhance the challenge voting 
page with QR code generation. Updates routing to include a new 
voting path and implements a Twig template for displaying the 
QR code, game PIN, and voting URL.
