+++ web/modules/custom/challenge/templates/challenge-voting-page.html.twig
@@ -1,0 +1,18 @@
+<article class="challenge-voting">
+  {% include 'wyciwyg:sdc_back' %}
+  <div class="content">
+    <h1>{{ label }}</h1>
+    <div class="grid"> 
+      <div class="qrcode">
+       <img src="{{ qr_code_uri }}" alt="">
+      </div>
+      <div class="url">
+       <span>wyciwyg.dev/vote</span>
+      </div>
+      <div class="gamepin">
+        <span>{{ game_pin }}</span>
+        <p>Game PIN</p>
+      </div>
+    </div>
+  </div>
+</article>
